Remove lens change documentation.
The lens can't be changed at runtime because there is no way to know
which lens was read in from the scene file.

One way around this would be to check the size of ss.lens as Ortho and
Perspective have a uniq number of vectors, and provide a default lens
only if that lens type is not provided.